+++ CHANGELOG.md
@@ -85,6 +85,7 @@
 - Fix trading name not being imported from opportunity on sale quotation generation
 - Purchase order: fix doubled stock move total when generating supplier arrival.
 - Logistical form is now printed in customer language and not in user language.
+- Timesheet editor: fix java.lang.NullPointerException in TimesheetServiceImpl when delete a line in editor
 
 ## [5.0.0-rc2] - 2018-05-09
 ## Improvements